requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
alembic==1.15.2
attrs==25.3.0
beautifulsoup4==4.14.2
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==46.0.3
dill==0.3.8
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.1.0
Flask-Caching==2.3.1
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.6.0
fsspec==2025.3.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
msal==1.34.0
multidict==6.4.4
multiprocess==0.70.16
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
propcache==0.3.1
pytest==8.3.5
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
python-utils==3.9.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.40
tqdm==4.67.1
typing_extensions==4.13.1
tzdata==2025.1
urllib3==2.3.0
waitress==3.0.2
Werkzeug==3.1.3
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
